+++ CHANGELOG.md
@@ -15,9 +15,9 @@
 - `f-drawer` 组件 `close-icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
 - 新增 `f-drawer` 组件 `closeIcon` 插槽
 - 新增 `f-close-btn` 组件 `color` 属性
+- `f-link` 组件 `before-icon` `after-icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
+- 新增 `f-link` 组件 `beforeIcon` `afterIcon` 插槽
 
-- 废除 `f-ink` 组件 `before-icon` `after-icon` 配置项
-- 新增 `f-ink` 组件 `beforeIcon` `afterIcon` 插槽
 - 废除 `f-avatar` 组件 `icon` 配置项
 - 新增 `f-avatar` 组件 `icon` 插槽
 - 废除 `f-loading` 组件 `icon` 配置项